Modified the way the address is recorded. A voter can have multiple addresses and we can diferentiate them by an address type. Because of this modification, had to modify the way the voter, person and address domains are saved; and also the way the voter information is loaded from a spreadsheet.
The voters list are no longer sortable because of the above changes. We
are not able to utilize the hibernate sessions because zk loses them, so
we are passing the data as maps to the view. The sortable feature will
have to wait for a future release.

The voters detail view is broken because of the change to the address
domain. I need to fix the address panels first in order to continue
working on the details views.